@@ -1,12 +1,4 @@
 <?php
-/**
- * Cache Flag plugin for Craft CMS 3.x
- *
- * Flag and clear template caches.
- *
- * @link      https://vaersaagod.no
- * @copyright Copyright (c) 2018 Mats Mikkel Rummelhoff
- */
 
 namespace mmikkel\cacheflag\controllers;
 
@@ -26,15 +18,15 @@ use yii\web\ForbiddenHttpException;
 class DefaultController extends Controller
 {
 
-    // Public Methods
-    // =========================================================================
+    /** @var array|bool|int */
+    public array|bool|int $allowAnonymous = self::ALLOW_ANONYMOUS_NEVER;
 
     /**
      * @return \yii\web\Response
      * @throws BadRequestHttpException
      * @throws \yii\base\InvalidConfigException
      */
-    public function actionSaveFlags()
+    public function actionSaveFlags(): \yii\web\Response
     {
         $this->requirePostRequest();
         $this->requireAcceptsJson();
@@ -85,21 +77,10 @@ class DefaultController extends Controller
 
         return $this->asJson([
             'success' => true,
-            'message' => $this->_getRandomSuccessMessage(),
+            'message' => Craft::t('cache-flag', 'Cache flags saved'),
             'flags' => CacheFlag::getInstance()->cacheFlag->getAllFlags(),
         ]);
 
-    }
-
-    /**
-     * @return \yii\web\Response
-     * @throws BadRequestHttpException
-     * @throws \yii\base\InvalidConfigException
-     * @deprecated since 1.1.0
-     */
-    public function actionDeleteFlaggedCachesByFlags()
-    {
-        return $this->actionInvalidateFlaggedCachesByFlags();
     }
 
     /**
@@ -139,19 +120,8 @@ class DefaultController extends Controller
 
         return $this->asJson([
             'success' => true,
-            'message' => Craft::t('cache-flag', 'Caches invalidated'),
+            'message' => Craft::t('cache-flag', 'Flagged caches invalidated'),
         ]);
-    }
-
-    /**
-     * @return \yii\web\Response
-     * @throws BadRequestHttpException
-     * @throws \craft\errors\MissingComponentException
-     * @deprecated since 1.1.0
-     */
-    public function actionDeleteAllFlaggedCaches()
-    {
-        return $this->actionInvalidateAllFlaggedCaches();
     }
 
     /**
@@ -178,54 +148,6 @@ class DefaultController extends Controller
         }
 
         return $this->redirectToPostedUrl();
-    }
-
-    /**
-     * @return string
-     */
-    private function _getRandomSuccessMessage(): string
-    {
-        $messages = [
-            "Flags saved! Good day to you.",
-            "Ahoy! All flags saved.",
-            "Flags. Saved. High five.",
-            "Flags! We haz them.",
-            "Cache flags saved, friend.",
-            "Yarr! Yer flags be safe.",
-            "Dun dun dun. All flags saved.",
-            "You are awesome. And so are your flags.",
-            "Flags, so many flags. All mine!",
-            "Flags flags flags flags",
-            "You say save, I say sure.",
-            "Before, your flags were lost. But now they are saved.",
-            "My, what beautiful flags you have.",
-            "Your flags are safe with me.",
-            "I'll just hang on to these flags for you.",
-            "If you want a symbolic gesture, don't burn the flag; save it.",
-            'Those are the second biggest flags I\'ve ever seen!',
-            'Put a flag on it.',
-            'No worries, your flags are totally saved.',
-            'Current state: flags saved.',
-            'Mhmm, yummy flags been saved.',
-            'Much flags. So saved.',
-            'Your Cache Flags (tm) are safely saved.',
-            'Flags saved! Go forth and conquer!',
-            'Saving flags is my favorite thing in the world',
-            'Safe and sound.',
-            'It\'s getting hot in here',
-            'Flags. Saved. Profit.',
-            'These might just be the prettiest lil\' flags I\'ve ever saved',
-            'Saving cache flags sure is fun, huh',
-            'Come on! Moar flags!',
-            'Thank you for these wonderful flags',
-            'Cache Flags! Exactly what I needed right now!',
-            'Nice one! Successfully saved your flags.',
-            'Oh man, just wait until people see these flags',
-            'Not gonna lie. Those are some fine flags.',
-            'Your flags are safe and sound.',
-            'Tweet it out?',
-        ];
-        return $messages[\array_rand($messages)];
     }
 
 }
